refactor(mcp): extract sse transport module
Part of #3310.

Move the legacy SSE transport, endpoint discovery loop, and POST/recv implementation out of the MCP monolith into mcp/sse.rs. Keep the HTTP fallback coordinator and streamable HTTP transport in mcp.rs so this remains a behavior-preserving split.

Verification:

- cargo fmt --all -- --check

- git diff --check

- CARGO_TARGET_DIR=/Volumes/VIXinSSD/codewhale-target cargo test -p codewhale-tui mcp::tests --locked

- CARGO_TARGET_DIR=/Volumes/VIXinSSD/codewhale-target cargo check -p codewhale-tui --bin codewhale-tui --locked
